add polygonPreprocessorSmokeTest function, reorganize smoke test utility methods
